Fix bug:
With one life left and one asteroid left,
if ship kamikaze's that asteroid, the game ends
yet bcuz that was the last asteroid, the level was increasing.

Check for ship-collision after the asteroids have been drawn
and consequently after the game determines if there aren't any asteroids
and increases the level.